pyproject.toml
[tool.poetry]
name = "spacerunner"
version = "0.1.0"
description = "A tool for running Maker/Hacker/Other spaces"
authors = ["Matthew Macdonald-Wallace <matt@doics.co>"]
license = "MIT-Modern-Variant"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
django-prometheus = "^2.2.0"
django-bootstrap4 = "^22.1"
django-mptt = "^0.13.4"
django-jazzmin = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
